# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: VSMS Event Operations API
  version: 2.4.0
  description: Secure staff authentication, event operations, and clinical review workflows for VSMS.
  license:
    name: ISC
    url: https://opensource.org/license/isc-license-txt
servers:
  - url: /
    description: Same-origin API server; local development proxies to https://127.0.0.1:5050
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/events:
    get:
      tags: [Events]
      operationId: listEvents
      summary: List events visible to the current staff member
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/EventLimit"
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/EventStatus" }
        - name: search
          in: query
          schema: { type: string, maxLength: 150 }
      responses:
        "200":
          description: Authorized, cursor-paginated events
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    post:
      tags: [Events]
      operationId: createEvent
      summary: Create a draft event with optional planned shifts
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Stable retry key for this create attempt
          schema: { type: string, minLength: 8, maxLength: 100, pattern: '^[A-Za-z0-9_-]+$' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateEventRequest" }
      responses:
        "201":
          description: Event created and audited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/events/active:
    get:
      tags: [Events]
      operationId: listActiveEvents
      summary: List published or in-progress events
      responses:
        "200":
          description: Active events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Event" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/events/station-templates:
    get:
      tags: [Events]
      operationId: listStationTemplates
      summary: List active reusable station templates
      responses:
        "200":
          description: Active station templates available for import
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StationTemplate" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/v1/public/events/{eventId}:
    get:
      tags: [Events]
      security: []
      operationId: getPublicEvent
      summary: Get public details for a non-draft event
      parameters: [{ $ref: "#/components/parameters/EventId" }]
      responses:
        "200":
          description: Public event detail
          headers:
            Cache-Control: { schema: { type: string, enum: ["public, max-age=60"] } }
          content: { application/json: { schema: { $ref: "#/components/schemas/PublicEvent" } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /api/v1/events/{eventId}/attendees:
    get:
      tags: [Events]
      operationId: listEventAttendees
      summary: List event attendees for an authorized manager
      parameters:
        - $ref: "#/components/parameters/EventId"
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: status
          in: query
          schema: { type: string, enum: [SIGNED_UP, CHECKED_IN, COMPLETED, CANCELLED] }
        - name: search
          in: query
          schema: { type: string, maxLength: 150 }
      responses:
        "200": { description: Attendee page, content: { application/json: { schema: { $ref: "#/components/schemas/EventAttendeeList" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/events/{eventId}/export:
    get:
      tags: [Events]
      operationId: exportEvent
      summary: Export an authorized event operations snapshot
      parameters: [{ $ref: "#/components/parameters/EventId" }]
      responses:
        "200": { description: Event export snapshot, content: { application/json: { schema: { $ref: "#/components/schemas/EventExportResponse" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/events/{eventId}:
    parameters:
      - $ref: "#/components/parameters/EventId"
    get:
      tags: [Events]
      operationId: getEvent
      summary: Get an authorized event detail
      responses:
        "200":
          description: Event detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Events]
      operationId: updateEvent
      summary: Update allowed fields using optimistic concurrency
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateEventRequest" }
      responses:
        "200":
          description: Event updated and audited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      tags: [Events]
      operationId: deleteTerminalEvent
      summary: Permanently delete a completed or cancelled event
      description: Administrator-only hard deletion. Event-owned operational records are removed; shared staff accounts, participants, templates, forms, and devices are retained.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventDeletionRequest" }
      responses:
        "200":
          description: Event permanently deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventDeletionResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
  /api/v1/events/{eventId}/cancel:
    post:
      tags: [Events]
      operationId: cancelEvent
      summary: Cancel a non-terminal event with a reason
      parameters: [{ $ref: "#/components/parameters/EventId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CancelEventRequest" }
      responses: { "200": { $ref: "#/components/responses/EventSuccess" }, "401": { $ref: "#/components/responses/Unauthorized" }, "403": { $ref: "#/components/responses/Forbidden" }, "404": { $ref: "#/components/responses/NotFound" }, "409": { $ref: "#/components/responses/Conflict" }, "422": { $ref: "#/components/responses/ValidationFailed" }, "429": { $ref: "#/components/responses/RateLimited" } }
...
  /api/v1/participants:
    get:
      tags: [Participants]
      operationId: searchParticipants
      summary: Search participants by at least one identifying field
      parameters:
        - { name: participantReference, in: query, schema: { type: string, minLength: 3, maxLength: 30 } }
        - { name: name, in: query, schema: { type: string, minLength: 3, maxLength: 100 } }
        - { name: contactNumber, in: query, schema: { type: string, minLength: 3, maxLength: 30 } }
        - { name: dateOfBirth, in: query, schema: { type: string, format: date } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/ParticipantPageSize" }
      responses:
        "200": { description: Paginated participant summaries, content: { application/json: { schema: { $ref: "#/components/schemas/ParticipantSearchResponse" } } } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
    post:
      tags: [Participants]
      operationId: createParticipant
      summary: Create a participant
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ParticipantRequest" } } } }
      responses:
        "201": { description: Participant created, content: { application/json: { schema: { $ref: "#/components/schemas/ParticipantEnvelope" } } } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/v1/participants/{participantId}:
    parameters: [{ $ref: "#/components/parameters/ParticipantId" }]
    get:
      tags: [Participants]
      operationId: getParticipant
      summary: Get a participant and emergency contacts
      responses:
        "200": { description: Participant detail, content: { application/json: { schema: { $ref: "#/components/schemas/ParticipantEnvelope" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Participants]
      operationId: updateParticipant
      summary: Update supplied participant fields
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ParticipantUpdateRequest" } } } }
      responses:
        "200": { description: Participant updated, content: { application/json: { schema: { $ref: "#/components/schemas/ParticipantEnvelope" } } } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/consent-forms/active:
    get:
      tags: [Participants]
      operationId: getActiveConsentForm
      summary: Get the currently effective consent form
      responses:
        "200": { description: Active consent form, content: { application/json: { schema: { $ref: "#/components/schemas/ConsentFormEnvelope" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/emergency-contacts/{contactId}:
    patch:
      tags: [Participants]
      operationId: updateEmergencyContact
      summary: Update an emergency contact by ID
      parameters: [{ $ref: "#/components/parameters/ContactId" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/EmergencyContactUpdateRequest" } } } }
      responses:
        "200": { description: Emergency contact updated, content: { application/json: { schema: { $ref: "#/components/schemas/EmergencyContactEnvelope" } } } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token kept in browser memory only
    refreshCookie:
      type: apiKey
      in: cookie
      name: vsms_refresh
  parameters:
    EventId:
      name: eventId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ParticipantId:
      name: participantId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ContactId:
      name: contactId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string, maxLength: 2048 }
    EventLimit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, maximum: 10000, default: 1 }
...
    ParticipantPageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
...
  responses:
    EventSuccess:
      description: Event transitioned and audited
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Event" }
...
    Unauthorized:
      description: Authentication is absent, invalid, expired, revoked, or reused
      content: { application/json: { schema: { $ref: "#/components/schemas/Problem" } } }
    Forbidden:
      description: Request origin, CSRF token, or permission is invalid
      content: { application/json: { schema: { $ref: "#/components/schemas/Problem" } } }
    NotFound:
      description: Resource is absent or intentionally concealed from the caller
      content: { application/json: { schema: { $ref: "#/components/schemas/Problem" } } }
    Conflict:
      description: Stale version, illegal transition, or immutable state
      content: { application/json: { schema: { $ref: "#/components/schemas/Problem" } } }
    ValidationFailed:
      description: Request validation failed
      content: { application/json: { schema: { $ref: "#/components/schemas/Problem" } } }
    RateLimited:
      description: Rate limit exceeded
      content: { application/json: { schema: { $ref: "#/components/schemas/Problem" } } }
...
  schemas:
...
    EventStatus:
      type: string
      enum: [DRAFT, PUBLISHED, IN_PROGRESS, COMPLETED, CANCELLED]
    EventBannerKey:
      type: string
      enum: [COMMUNITY_SCREENING, LIBRARY_SCREENING, EVENT_OPERATIONS]
...
    StationTemplate:
      type: object
      additionalProperties: false
      required: [stationTemplateId, templateKey, version, name, defaultCapacity]
      properties:
        stationTemplateId: { type: string, format: uuid }
        templateKey: { type: string, maxLength: 80 }
        version: { type: integer, minimum: 1 }
        name: { type: string, maxLength: 120 }
        description: { type: string, nullable: true }
        defaultCapacity: { type: integer, minimum: 1, maximum: 1000 }
...
    EventDayInput:
      type: object
      additionalProperties: false
      required: [date, startsAt, endsAt]
      properties:
        eventDayId: { type: string, format: uuid }
        date: { type: string, format: date }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
    EventDay:
      allOf:
        - $ref: "#/components/schemas/EventDayInput"
        - type: object
          required: [eventDayId]
          properties:
            eventDayId: { type: string, format: uuid }
...
    EventStationInput:
      type: object
      additionalProperties: false
      required: [stationTemplateId, stationOrder, capacity, isAvailable, availabilities]
      properties:
        eventStationId: { type: string, format: uuid }
        stationTemplateId: { type: string, format: uuid }
        stationOrder: { type: integer, minimum: 1, maximum: 50 }
        capacity: { type: integer, minimum: 1, maximum: 100000 }
        isAvailable: { type: boolean }
        availabilities:
          type: array
          maxItems: 31
          items: { $ref: "#/components/schemas/StationAvailabilityInput" }
...
    ShiftInput:
      type: object
      additionalProperties: false
      required: [name, startsAt, endsAt, requiredStaff]
      properties:
        shiftId: { type: string, format: uuid, description: Required only when editing an existing shift }
        name: { type: string, minLength: 1, maxLength: 100 }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        requiredStaff: { type: integer, minimum: 1, maximum: 1000, default: 1 }
        assignments:
          type: array
          maxItems: 100
          items: { $ref: "#/components/schemas/ShiftAssignmentInput" }
    Shift:
      allOf:
        - $ref: "#/components/schemas/ShiftInput"
        - type: object
          required: [shiftId, status, staffAssignments]
          properties:
            shiftId: { type: string, format: uuid }
            status: { $ref: "#/components/schemas/ShiftStatus" }
            staffAssignments: { type: array, items: { $ref: "#/components/schemas/StaffAssignment" } }
    CreateEventRequest:
      type: object
      additionalProperties: false
      required: [name, venue, timezone, startsAt, endsAt, capacity]
      properties:
        name: { type: string, minLength: 1, maxLength: 150 }
        description: { type: string, nullable: true, maxLength: 5000 }
        bannerKey: { $ref: "#/components/schemas/EventBannerKey" }
        artworkDataUrl:
          type: string
          nullable: true
          maxLength: 180000
          pattern: '^data:image/(jpeg|webp);base64,[A-Za-z0-9+/]+={0,2}$'
          description: Cropped square event artwork encoded as a bounded JPEG or WebP data URL
        venue: { type: string, minLength: 1, maxLength: 255 }
        address: { type: string, nullable: true, maxLength: 500 }
        postalCode: { type: string, nullable: true, pattern: '^\d{6}$' }
        latitude: { type: number, format: double, nullable: true, minimum: 1.13, maximum: 1.48 }
        longitude: { type: number, format: double, nullable: true, minimum: 103.59, maximum: 104.1 }
        locationProvider: { type: string, nullable: true, enum: [ONEMAP, MANUAL] }
        locationReference: { type: string, nullable: true, maxLength: 255 }
        timezone: { type: string, minLength: 1, maxLength: 100, example: Asia/Singapore }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        capacity: { type: integer, minimum: 1, maximum: 100000 }
        expectedAttendance: { type: integer, nullable: true, minimum: 1, maximum: 1000000 }
        eventDays: { type: array, minItems: 1, maxItems: 31, items: { $ref: "#/components/schemas/EventDayInput" } }
        stations: { type: array, maxItems: 50, items: { $ref: "#/components/schemas/EventStationInput" } }
        shifts: { type: array, maxItems: 50, items: { $ref: "#/components/schemas/ShiftInput" } }
    UpdateEventRequest:
      type: object
      additionalProperties: false
      required: [version]
      properties:
        version: { type: integer, minimum: 1 }
        name: { type: string, minLength: 1, maxLength: 150 }
        description: { type: string, nullable: true, maxLength: 5000 }
        bannerKey: { $ref: "#/components/schemas/EventBannerKey" }
        artworkDataUrl:
          type: string
          nullable: true
          maxLength: 180000
          pattern: '^data:image/(jpeg|webp);base64,[A-Za-z0-9+/]+={0,2}$'
        venue: { type: string, minLength: 1, maxLength: 255 }
        address: { type: string, nullable: true, maxLength: 500 }
        postalCode: { type: string, nullable: true, pattern: '^\d{6}$' }
        latitude: { type: number, format: double, nullable: true }
        longitude: { type: number, format: double, nullable: true }
        locationProvider: { type: string, nullable: true, enum: [ONEMAP, MANUAL] }
        locationReference: { type: string, nullable: true, maxLength: 255 }
        timezone: { type: string, minLength: 1, maxLength: 100 }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        capacity: { type: integer, minimum: 1, maximum: 100000 }
        expectedAttendance: { type: integer, nullable: true, minimum: 1, maximum: 1000000 }
        eventDays: { type: array, minItems: 1, maxItems: 31, items: { $ref: "#/components/schemas/EventDayInput" } }
        stations: { type: array, maxItems: 50, items: { $ref: "#/components/schemas/EventStationInput" } }
        shifts: { type: array, maxItems: 50, items: { $ref: "#/components/schemas/ShiftInput" } }
    CancelEventRequest:
      type: object
      additionalProperties: false
      required: [version, reason]
      properties:
        version: { type: integer, minimum: 1 }
        reason: { type: string, minLength: 10, maxLength: 1000 }
    EventDeletionRequest:
      type: object
      additionalProperties: false
      required: [version, confirmationName, acknowledgePermanentDeletion]
      properties:
        version: { type: integer, minimum: 1 }
        confirmationName: { type: string, minLength: 1, maxLength: 150, description: Exact event name acknowledgement }
        acknowledgePermanentDeletion: { type: boolean, enum: [true] }
    EventDeletionResponse:
      type: object
      additionalProperties: false
      required: [eventId, deleted]
      properties:
        eventId: { type: string, format: uuid }
        deleted: { type: boolean, enum: [true] }
    Event:
      allOf:
        - $ref: "#/components/schemas/CreateEventRequest"
        - type: object
          required: [eventId, artworkDataUrl, status, version, shifts, eventDays, eventStations, signupCount, activeCapacityCount, canManage, createdAt, updatedAt]
          properties:
            eventId: { type: string, format: uuid }
            status: { $ref: "#/components/schemas/EventStatus" }
            version: { type: integer, minimum: 1 }
            cancelledAt: { type: string, format: date-time, nullable: true }
            cancellationReason: { type: string, nullable: true }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
            shifts:
              type: array
              description: Managers receive the staffing plan; operational staff receive only their own ASSIGNED or CONFIRMED assignments on PLANNED or ACTIVE shifts.
              items: { $ref: "#/components/schemas/Shift" }
            eventDays: { type: array, items: { $ref: "#/components/schemas/EventDay" } }
            eventStations: { type: array, items: { $ref: "#/components/schemas/EventStation" } }
            signupCount: { type: integer, minimum: 0, description: Number of collected registration entries for this event }
            activeCapacityCount: { type: integer, minimum: 0, description: Number of registrations checked in to the event; this is not a live count of people still at the venue }
            canManage: { type: boolean, description: Whether the current caller may change this event }
            createdBy:
              description: Manager-only organisation identity projection.
              allOf: [{ $ref: "#/components/schemas/User" }]
            cancelledBy:
              description: Manager-only organisation identity projection.
              type: object
              allOf: [{ $ref: "#/components/schemas/User" }]
              nullable: true
    PublicEvent:
      type: object
      additionalProperties: false
      required: [eventId, name, bannerKey, venue, timezone, startsAt, endsAt, capacity, status, eventDays]
      properties:
        eventId: { type: string, format: uuid }
        name: { type: string, maxLength: 150 }
        description: { type: string, nullable: true }
        bannerKey: { type: string }
        artworkDataUrl: { type: string, nullable: true }
        venue: { type: string }
        address: { type: string, nullable: true }
        postalCode: { type: string, nullable: true }
        timezone: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        capacity: { type: integer, minimum: 1 }
        status: { $ref: "#/components/schemas/EventStatus" }
        eventDays: { type: array, items: { $ref: "#/components/schemas/EventDay" } }
...
    EventAttendeeList:
      type: object
      additionalProperties: false
      required: [total, attendees, nextCursor]
      properties:
        total: { type: integer, minimum: 0 }
        attendees: { type: array, items: { $ref: "#/components/schemas/EventAttendee" } }
        nextCursor: { type: string, nullable: true }
    EventExportResponse:
      type: object
      additionalProperties: false
      required: [export]
      properties:
        export:
          type: object
          additionalProperties: true
        exportReceipt:
          type: string
...
    EventListResponse:
      type: object
      required: [events, nextCursor]
      properties:
        events: { type: array, items: { $ref: "#/components/schemas/Event" } }
        nextCursor: { type: string, nullable: true }
...
    ParticipantRequest:
      type: object
      additionalProperties: false
      required: [firstName, lastName, dateOfBirth, gender, contactNumber]
      properties:
        firstName: { type: string, minLength: 1, maxLength: 100 }
        lastName: { type: string, minLength: 1, maxLength: 100 }
        dateOfBirth: { type: string, format: date }
        gender: { type: string, enum: [M, F, O, U] }
        contactNumber: { type: string, minLength: 7, maxLength: 30 }
        email: { type: string, format: email, nullable: true }
        preferredLanguage: { type: string, maxLength: 50, nullable: true }
        accessibilityNotes: { type: string, maxLength: 1000, nullable: true }
        status: { type: string, enum: [ACTIVE, INACTIVE, DECEASED] }
    ParticipantUpdateRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        firstName: { type: string, minLength: 1, maxLength: 100 }
        lastName: { type: string, minLength: 1, maxLength: 100 }
        dateOfBirth: { type: string, format: date }
        gender: { type: string, enum: [M, F, O, U] }
        contactNumber: { type: string, minLength: 7, maxLength: 30 }
        email: { type: string, format: email, nullable: true }
        preferredLanguage: { type: string, maxLength: 50, nullable: true }
        accessibilityNotes: { type: string, maxLength: 1000, nullable: true }
        status: { type: string, enum: [ACTIVE, INACTIVE, DECEASED] }
...
    ParticipantSummary:
      type: object
      additionalProperties: false
      required: [id, participantReference, firstName, lastName, maskedContactNumber, maskedDateOfBirth, status]
      properties:
        id: { type: string, format: uuid }
        participantReference: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        maskedContactNumber: { type: string }
        maskedDateOfBirth: { type: string }
        status: { type: string, enum: [ACTIVE, INACTIVE, DECEASED] }
    ParticipantEnvelope:
      type: object
      additionalProperties: false
      required: [participant]
      properties: { participant: { $ref: "#/components/schemas/Participant" } }
    ParticipantSearchResponse:
      type: object
      additionalProperties: false
      required: [participants, pagination]
      properties:
        participants: { type: array, items: { $ref: "#/components/schemas/ParticipantSummary" } }
        pagination: { $ref: "#/components/schemas/Pagination" }
...
    EmergencyContactUpdateRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        contactName: { type: string, minLength: 1, maxLength: 120 }
        relationship: { type: string, minLength: 1, maxLength: 60 }
        phoneNumber: { type: string, minLength: 7, maxLength: 30 }
        email: { type: string, format: email, nullable: true }
        isPrimary: { type: boolean }
        status: { type: string, enum: [ACTIVE, REMOVED] }
...
    EmergencyContactEnvelope:
      type: object
      additionalProperties: false
      required: [contact]
      properties: { contact: { $ref: "#/components/schemas/EmergencyContact" } }
...
    ConsentFormEnvelope:
      type: object
      additionalProperties: false
      required: [consentForm]
      properties: { consentForm: { $ref: "#/components/schemas/ConsentForm" } }
...
    Problem:
      type: object
      required: [type, title, status, code, requestId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        requestId: { type: string, format: uuid }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }